requirements.txt
asgiref==3.10.0; python_version >= '3.9'
astroid==3.3.11; python_full_version >= '3.9.0'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
dill==0.4.0; python_version >= '3.8'
django==5.2.7; python_version >= '3.10'
django-debug-toolbar==6.0.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
isort==6.1.0; python_full_version >= '3.9.0'
kafka-python==2.2.15
mccabe==0.7.0; python_version >= '3.6'
pillow==12.0.0; python_version >= '3.10'
platformdirs==4.5.0; python_version >= '3.10'
psycopg2-binary==2.9.11; python_version >= '3.9'
pylint==3.3.9; python_full_version >= '3.9.0'
pylint-django==2.6.1; python_version >= '3.9' and python_version < '4.0'
pylint-plugin-utils==0.9.0; python_version >= '3.9' and python_version < '4.0'
python-decouple==3.8
sqlparse==0.5.3; python_version >= '3.8'
tomlkit==0.13.3; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
confluent-kafka==2.12.1; python_version >= '3.7'